feat: Completed work on dspace repository.
Dspace results now appear in the cross search repository reponse.  However, it appears that DSpace solr does not return the parent collection name.  It does return the collection id and the dc.relation.ispartofseries (if this was added...some librarians been doing so and I noticed inconsistency in naming across collections).

Issues: AC-1253