feat: Modifed response to return map of results with the collection name as the key.
Also added dspace collection name lookup in the application configuration. Listerner attached to the ApplicationReadyEvent.  It triggers a query of the dspace REST api and creates an in memory map of key value pairs for the collection ID and collection name.  This is used to add the collection name to dspace solr responses.  Unfortunately, dspace security prevents access to all collections via the REST api.  We need to explore more options. (Perhaps querying postgres directly.) Also a few changes to handle errors.

Issues: AC-1256, AC-1256